backend/requirements-full.txt
fastapi
uvicorn
pydantic
scikit-learn
pandas
numpy
python-multipart
tensorflow
pillow
python-jose[cryptography]
passlib[bcrypt]
